Convert the character literal syntax back to the conventional 'x'.
Talked to Arya, sounds like agreement has been reached that this scans
much better than ?x.

This is still WIP; in particular the patch as-is introduces a bug
because '( 'bar) gets printed as '('bar), which is now not equivalent.
Need to adjust the printer to add a space where necessary.